add: add definition
