--- src/os_sanic/workflow.py
@@ -1,9 +1,0 @@
-from abc import ABCMeta
-
-
-class Workflowable(metaclass=ABCMeta):
-    def setup(self):
-        pass
-
-    def cleanup(self):
-        pass
